changes in backend folder structure
