+++ backend/.gitignore
@@ -1,7 +1,9 @@
 # Ignore node_modules directory
 node_modules/
+node_modules
 
 # Ignore .env file
 .env
 
 backup/
+backup
